- Several bugfixes for creation process. - Version 2.05, instead of 2.1 as structural edit for massletter is not being worked on. - Work for disabling delete of users on demo install. (Should probably also disable change of password) - Some more work on StructuralTableEdit. - Support for direct link login to be used for demo.
git-svn-id: https://accountserver.googlecode.com/svn/trunk@309 7b885804-ab2d-0410-b250-0b2d121ae3f5

+++ RegnskapServer/services/registers/users.php
@@ -4,6 +4,7 @@ include_once ("../../classes/util/DB.php");
 include_once ("../../classes/auth/User.php");
 include_once ("../../classes/auth/RegnSession.php");
 include_once ("../../classes/auth/Master.php");
+include_once ("../../classes/accounting/accountstandard.php");
 
 $action = array_key_exists("action", $_REQUEST) ? $_REQUEST["action"] : "all";
 $user = array_key_exists("username", $_REQUEST) ? $_REQUEST["username"] : "";
@@ -64,6 +65,16 @@ switch ($action) {
         break;
     case "delete" :
         $regnSession->checkWriteAccess();
+        
+        $accStandard = new AccountStandard($db);
+
+        $abortDelete = $accStandard->getOneValue(AccountStandard::CONST_NO_DELETE_USERS);
+        
+        if($abortDelete) {
+            header("HTTP/1.0 513 Validation Error");
+            die(json_encode(array("DELETE_DISABLED")));
+        }
+        
         $accUsers = new User($db);
         $rowsAffected = $accUsers->delete($user);
         $res = array ();
